# docs/swagger.yaml
basePath: /
definitions:
  api.AVstream:
    properties:
      aqueue:
        format: uint64
        type: integer
      channels:
        format: uint64
        type: integer
      codec:
        description: Codec parameter
        type: string
      drop:
        format: uint64
        type: integer
      dup:
        format: uint64
        type: integer
      duplicating:
        type: boolean
      enc:
        format: uint64
        type: integer
      gop:
        type: string
      height:
        format: uint64
        type: integer
      input:
        $ref: '#/definitions/api.AVstreamIO'
      layout:
        type: string
      level:
        type: integer
      looping:
        type: boolean
      looping_runtime:
        format: uint64
        type: integer
      mode:
        type: string
      output:
        $ref: '#/definitions/api.AVstreamIO'
      pix_fmt:
        type: string
      profile:
        type: integer
      queue:
        format: uint64
        type: integer
      sample_fmt:
        type: string
      sampling_hz:
        format: uint64
        type: integer
      width:
        format: uint64
        type: integer
    type: object
...
  api.ClusterKVS:
    additionalProperties:
      $ref: '#/definitions/api.ClusterKVSValue'
    type: object
  api.ClusterKVSValue:
    properties:
      updated_at:
        type: string
      value:
        type: string
    type: object
  api.ClusterLock:
    properties:
      name:
        type: string
      valid_until:
        type: string
    type: object
...
  api.ClusterProcessMap:
    additionalProperties:
      type: string
    type: object
...
  api.ClusterProcessRelocateMap:
    additionalProperties:
      type: string
    type: object
...
  api.ClusterStoreNode:
    properties:
      id:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
...
  api.ConfigError:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
  api.GetConfig:
    properties:
      config:
        $ref: '#/definitions/api.ConfigData'
      created_at:
        type: string
      loaded_at:
        type: string
      overrides:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  api.GraphMapping:
    properties:
      copy:
        type: boolean
      id:
        type: string
      index:
        type: integer
      input:
        type: integer
      name:
        type: string
      output:
        type: integer
    type: object
...
  api.IAMPolicy:
    properties:
      actions:
        items:
          type: string
        type: array
      domain:
        type: string
      name:
        type: string
      resource:
        type: string
      types:
        items:
          type: string
        type: array
    type: object
  api.IAMUser:
    properties:
      alias:
        type: string
      auth:
        $ref: '#/definitions/api.IAMUserAuth'
      created_at:
        format: int64
        type: integer
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/api.IAMPolicy'
        type: array
      superuser:
        type: boolean
      updated_at:
        format: int64
        type: integer
    type: object
  api.IAMUserAuth:
    properties:
      api:
        $ref: '#/definitions/api.IAMUserAuthAPI'
      services:
        $ref: '#/definitions/api.IAMUserAuthServices'
    type: object
  api.IAMUserAuthAPI:
    properties:
      auth0:
        $ref: '#/definitions/api.IAMUserAuthAPIAuth0'
      userpass:
        type: string
    type: object
  api.IAMUserAuthAPIAuth0:
    properties:
      tenant:
        $ref: '#/definitions/api.IAMAuth0Tenant'
      user:
        type: string
    type: object
  api.IAMUserAuthServices:
    properties:
      basic:
        items:
          type: string
        type: array
      session:
        items:
          type: string
        type: array
      token:
        items:
          type: string
        type: array
    type: object
...
  api.Process:
    properties:
      config:
        $ref: '#/definitions/api.ProcessConfig'
      core_id:
        type: string
      created_at:
        format: int64
        type: integer
      domain:
        type: string
      id:
        type: string
      metadata: {}
      owner:
        type: string
      reference:
        type: string
      report:
        $ref: '#/definitions/api.ProcessReport'
      state:
        $ref: '#/definitions/api.ProcessState'
      type:
        type: string
      updated_at:
        format: int64
        type: integer
    type: object
  api.ProcessConfig:
    properties:
      autostart:
        type: boolean
      binary:
        type: string
      domain:
        type: string
      id:
        type: string
      input:
        items:
          $ref: '#/definitions/api.ProcessConfigIO'
        type: array
      limits:
        $ref: '#/definitions/api.ProcessConfigLimits'
      log_patterns:
        items:
          type: string
        type: array
      metadata:
        additionalProperties: true
        type: object
      options:
        items:
          type: string
        type: array
      output:
        items:
          $ref: '#/definitions/api.ProcessConfigIO'
        type: array
      owner:
        type: string
      reconnect:
        type: boolean
      reconnect_delay_seconds:
        format: uint64
        type: integer
      reference:
        type: string
      runtime_duration_seconds:
        format: uint64
        type: integer
      scheduler:
        type: string
      stale_timeout_seconds:
        format: uint64
        type: integer
      type:
        enum:
        - ffmpeg
        - ""
        type: string
    required:
    - input
    - output
    type: object
  api.ProcessConfigIO:
    properties:
      address:
        type: string
      cleanup:
        items:
          $ref: '#/definitions/api.ProcessConfigIOCleanup'
        type: array
      id:
        type: string
      options:
        items:
          type: string
        type: array
    required:
    - address
    type: object
  api.ProcessConfigIOCleanup:
    properties:
      max_file_age_seconds:
        format: uint
        type: integer
      max_files:
        format: uint
        type: integer
      pattern:
        type: string
      purge_on_delete:
        type: boolean
    required:
    - pattern
    type: object
  api.ProcessConfigLimits:
    properties:
      cpu_usage:
        description: percent 0-100*ncpu
        type: number
      gpu_decoder:
        description: percent 0-100
        type: number
      gpu_encoder:
        description: percent 0-100
        type: number
      gpu_memory_mbytes:
        description: megabytes
        format: uint64
        type: integer
      gpu_usage:
        description: percent 0-100
        type: number
      log_event_rate:
        description: rate limit for log events
        format: float64
        type: number
      memory_mbytes:
        description: megabytes
        format: uint64
        type: integer
      waitfor_seconds:
        description: seconds
        format: uint64
        type: integer
    type: object
...
  api.ProcessReport:
    properties:
      created_at:
        format: int64
        type: integer
      history:
        items:
          $ref: '#/definitions/api.ProcessReportHistoryEntry'
        type: array
      lines:
        additionalProperties:
          format: int64
          type: integer
        type: object
      log:
        items:
          items:
            type: string
          type: array
        type: array
      matches:
        items:
          type: string
        type: array
      prelude:
        items:
          type: string
        type: array
    type: object
  api.ProcessReportHistoryEntry:
    properties:
      created_at:
        format: int64
        type: integer
      exit_state:
        type: string
      exited_at:
        format: int64
        type: integer
      lines:
        additionalProperties:
          format: int64
          type: integer
        type: object
      log:
        items:
          items:
            type: string
          type: array
        type: array
      matches:
        items:
          type: string
        type: array
      prelude:
        items:
          type: string
        type: array
      progress:
        $ref: '#/definitions/api.Progress'
      resources:
        $ref: '#/definitions/api.ProcessUsage'
    type: object
...
  api.ProcessState:
    properties:
      command:
        items:
          type: string
        type: array
      cpu_usage:
        description: deprecated, use Resources.Memory.Current
        type: number
      exec:
        type: string
      last_logline:
        type: string
      limit_mode:
        type: string
      memory_bytes:
        description: deprecated, use Resources.CPU.Current
        format: uint64
        type: integer
      order:
        type: string
      pid:
        format: int32
        type: integer
      progress:
        $ref: '#/definitions/api.Progress'
      reconnect_seconds:
        format: int64
        type: integer
      resources:
        $ref: '#/definitions/api.ProcessUsage'
      runtime_seconds:
        format: int64
        type: integer
    type: object
  api.ProcessUsage:
    properties:
      cpu_usage:
        $ref: '#/definitions/api.ProcessUsageCPU'
      gpu:
        $ref: '#/definitions/api.ProcessUsageGPU'
      memory_bytes:
        $ref: '#/definitions/api.ProcessUsageMemory'
    type: object
  api.ProcessUsageCPU:
    properties:
      avg:
        type: number
      cur:
        type: number
      limit:
        type: number
      max:
        type: number
      ncpu:
        type: number
      throttling:
        type: boolean
    type: object
  api.ProcessUsageGPU:
    properties:
      decoder:
        $ref: '#/definitions/api.ProcessUsageGPUUsage'
      encoder:
        $ref: '#/definitions/api.ProcessUsageGPUUsage'
      index:
        type: integer
      memory_bytes:
        $ref: '#/definitions/api.ProcessUsageGPUMemory'
      usage:
        $ref: '#/definitions/api.ProcessUsageGPUUsage'
    type: object
...
  api.ProcessUsageMemory:
    properties:
      avg:
        type: number
      cur:
        format: uint64
        type: integer
      limit:
        format: uint64
        type: integer
      max:
        format: uint64
        type: integer
    type: object
  api.Progress:
    properties:
      bitrate_kbit:
        description: kbit/s
        type: number
      drop:
        format: uint64
        type: integer
      dup:
        format: uint64
        type: integer
      fps:
        type: number
      frame:
        format: uint64
        type: integer
      inputs:
        items:
          $ref: '#/definitions/api.ProgressIO'
        type: array
      mapping:
        $ref: '#/definitions/api.StreamMapping'
      outputs:
        items:
          $ref: '#/definitions/api.ProgressIO'
        type: array
      packet:
        format: uint64
        type: integer
      q:
        type: number
      size_kb:
        description: kbytes
        format: uint64
        type: integer
      speed:
        type: number
      started:
        type: boolean
      time:
        type: number
    type: object
  api.ProgressIO:
    properties:
      address:
        type: string
      avstream:
        allOf:
        - $ref: '#/definitions/api.AVstream'
        description: avstream
      bitrate_kbit:
        description: kbit/s
        type: number
      channels:
        format: uint64
        type: integer
      codec:
        type: string
      coder:
        type: string
      extradata_size_bytes:
        description: bytes
        format: uint64
        type: integer
      format:
        type: string
      fps:
        type: number
      frame:
        format: uint64
        type: integer
      framerate:
        $ref: '#/definitions/api.ProgressIOFramerate'
      height:
        format: uint64
        type: integer
      id:
        type: string
      index:
        description: General
        format: uint64
        type: integer
      keyframe:
        format: uint64
        type: integer
      layout:
        type: string
      level:
        type: integer
      packet:
        format: uint64
        type: integer
      pix_fmt:
        description: Video
        type: string
      pps:
        type: number
      profile:
        type: integer
      q:
        type: number
      sample_fmt:
        description: Audio
        type: string
      sampling_hz:
        format: uint64
        type: integer
      size_kb:
        description: kbytes
        format: uint64
        type: integer
      stream:
        format: uint64
        type: integer
      tee:
        description: Format specific
        items:
          $ref: '#/definitions/api.ProgressIOTee'
        type: array
      type:
        type: string
      width:
        format: uint64
        type: integer
    type: object
  api.ProgressIOFramerate:
    properties:
      avg:
        type: number
      max:
        type: number
      min:
        type: number
    type: object
  api.ProgressIOTee:
    properties:
      address:
        type: string
      fifo_enabled:
        type: boolean
      fifo_recovery_attempts_total:
        type: number
      fifo_state:
        type: string
      format:
        type: string
      id:
        type: string
      state:
        type: string
    type: object
...
  api.StreamMapping:
    properties:
      graphs:
        items:
          $ref: '#/definitions/api.GraphElement'
        type: array
      mapping:
        items:
          $ref: '#/definitions/api.GraphMapping'
        type: array
    type: object
...
  errors.Error:
    properties:
      code:
        format: int
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
info:
  contact:
    email: hello@datarhei.com
    name: datarhei Core Support
    url: https://www.datarhei.com
  description: Expose REST API for the datarhei Core
  license:
    name: Apache 2.0
    url: https://github.com/datarhei/core/v16/blob/main/LICENSE
  title: datarhei Core API
  version: "3.0"
paths:
...
  /api/v3/cluster/db/kv:
    get:
      description: List of KV in the cluster DB
      operationId: cluster-3-db-list-kv
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClusterKVS'
      security:
      - ApiKeyAuth: []
      summary: List KV in the cluster DB
      tags:
      - v16.?.?
  /api/v3/cluster/db/locks:
    get:
      description: List of locks in the cluster DB
      operationId: cluster-3-db-list-locks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ClusterLock'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List locks in the cluster DB
      tags:
      - v16.?.?
  /api/v3/cluster/db/map/process:
    get:
      description: Retrieve a map of which process is running on which node
      operationId: cluster-3-db-process-node-map
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClusterProcessMap'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a map of which process is running on which node
      tags:
      - v16.?.?
  /api/v3/cluster/db/map/reallocate:
    get:
      description: Retrieve a map of which processes should be relocated
      operationId: cluster-3-db-process-relocate-map
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClusterProcessRelocateMap'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a map of which processes should be relocated
      tags:
      - v16.?.?
  /api/v3/cluster/db/node:
    get:
      description: List of nodes in the cluster DB
      operationId: cluster-3-db-list-nodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ClusterStoreNode'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List nodes in the cluster DB
      tags:
      - v16.?.?
  /api/v3/cluster/db/policies:
    get:
      description: List of policies in the cluster
      operationId: cluster-3-db-list-policies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.IAMPolicy'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List of policies in the cluster
      tags:
      - v16.?.?
  /api/v3/cluster/db/process:
    get:
      description: List of processes in the cluster DB
      operationId: cluster-3-db-list-processes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Process'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List of processes in the cluster DB
      tags:
      - v16.?.?
  /api/v3/cluster/db/process/:id:
    get:
      description: Get a process in the cluster DB
      operationId: cluster-3-db-get-process
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Domain to act on
        in: query
        name: domain
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Process'
      security:
      - ApiKeyAuth: []
      summary: Get a process in the cluster DB
      tags:
      - v16.?.?
  /api/v3/cluster/db/user:
    get:
      description: List of identities in the cluster
      operationId: cluster-3-db-list-identities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.IAMUser'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List of identities in the cluster
      tags:
      - v16.?.?
  /api/v3/cluster/db/user/{name}:
    get:
      description: List of identities in the cluster
      operationId: cluster-3-db-get-identity
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.IAMUser'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: List of identities in the cluster
      tags:
      - v16.?.?
...
  /api/v3/config:
    get:
      description: Retrieve the currently active Restreamer configuration
      operationId: config-3-get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetConfig'
      security:
      - ApiKeyAuth: []
      summary: Retrieve the currently active Restreamer configuration
      tags:
      - v16.7.2
    put:
      consumes:
      - application/json
      description: Update the current Restreamer configuration by providing a complete
        or partial configuration. Fields that are not provided will not be changed.
      operationId: config-3-set
      parameters:
      - description: Restreamer configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/api.SetConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ConfigError'
      security:
      - ApiKeyAuth: []
      summary: Update the current Restreamer configuration
      tags:
      - v16.7.2
  /api/v3/config/reload:
    get:
      description: Reload the currently active configuration. This will trigger a
        restart of the Core.
      operationId: config-3-reload
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Reload the currently active configuration
      tags:
      - v16.7.2
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiRefreshKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  Auth0KeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"